requirements.txt
aws-sdk-bedrock-runtime
smithy-aws-core
pyaudio
rx
python-dotenv
boto3
